+++ outputconfig/output.go
@@ -246,6 +246,7 @@ type yamlTLS struct {
 	CA               string `yaml:"ca"`
 	Cert             string `yaml:"cert"`
 	Key              string `yaml:"key"`
+	KeyPassword      string `yaml:"key_password"`
 	AllowTLS12       bool   `yaml:"allow_tls12"`
 	AllowWeakCiphers bool   `yaml:"allow_weak_ciphers"`
 }