+++ webhook/register.go
@@ -91,6 +91,7 @@ type yamlTLS struct {
 	CA               string `yaml:"ca"`
 	Cert             string `yaml:"cert"`
 	Key              string `yaml:"key"`
+	KeyPassword      string `yaml:"key_password"`
 	AllowTLS12       bool   `yaml:"allow_tls12"`
 	AllowWeakCiphers bool   `yaml:"allow_weak_ciphers"`
 }
@@ -172,6 +173,9 @@ func buildOutput(name string, rawConfig []byte, coreMetrics audit.Metrics, om au
 		cfg.TLSCA = yc.TLS.CA
 		cfg.TLSCert = yc.TLS.Cert
 		cfg.TLSKey = yc.TLS.Key
+		if yc.TLS.KeyPassword != "" {
+			cfg.TLSKeyPassword = []byte(yc.TLS.KeyPassword)
+		}
 		cfg.TLSPolicy = &audit.TLSPolicy{
 			AllowTLS12:       yc.TLS.AllowTLS12,
 			AllowWeakCiphers: yc.TLS.AllowWeakCiphers,